Add wait for config map integration test helper
Also split big file into smaller ones.

[#167366400](https://www.pivotaltracker.com/story/show/167366400)
